Changed the TankTrack physics Linear and Angular damping to reduce slide and rotation momentum. Refactored TankTrack SetThrottle to simply set the new CurrentThrottle member. New DriveTrack() function actually applies the force to the TankTrack. Removed TankTrack.TickComponent, DriveTrack and ApplySidewaysForce functions are now called from the new OnHit function.  This effectively fixes the issue in which forces were being applied to the TankTrack while the Tank was in the air.
NOTE: THIS IS THE COMMIT IN WHICH THE TANK MOVMENT DIVERGED FROM WHAT I WOULD LIKE TO START FROM FOR A HOVER TANK GAME.
It diverged primarily due to the linear damping and the amount of sideways force being applied in ApplySidewaysForce.
The angular damping should probably be not as stringent, but shouldn't be 0 like it was before.
We will want to tune the linear damping and sideways force.
Lecture 182